add extraction analysis strategy
