+++ var/ramble/repos/builtin/base_classes/application-base/base_class.py
@@ -16,7 +16,6 @@ import re
 import shlex
 import shutil
 import stat
-import string
 import time
 from html import escape
 from typing import Dict, List
@@ -3464,280 +3463,11 @@ class ApplicationBase(ObjectMixin, metaclass=ApplicationMeta):
         Success criteria are defined within the application.py, but can also be
         injected in a workspace config.
         """
+        import ramble.analysis
 
-        if (
-            self.get_status() == ExperimentStatus.UNKNOWN
-            and not workspace.dry_run
-        ):
-            logger.warn(
-                f"Experiment has status {self.get_status()}. Skipping analysis..\n"
-            )
-            self.result.finalize(workspace)
-            return
-
-        def format_context(context_match, context_format):
-
-            context_val = {}
-            if isinstance(context_format, str):
-                for group in string.Formatter().parse(context_format):
-                    if group[1]:
-                        context_val[group[1]] = context_match[group[1]]
-
-            context_string = context_format.format(**context_val)
-            return context_string
-
-        # Exit early if read from cache works.
-        if self.result.read_cache(workspace, self):
-            self.result.finalize(workspace)
-            return
-
-        criteria_list = self.success_list
-        if not criteria_list:
-            criteria_list = ramble.success_criteria.ScopedCriteriaList()
-        criteria_list.reset()
-
-        files, f_defs, inmem_defs = self.analysis_dicts(criteria_list)
-
-        exp_lock = self.experiment_lock
-
-        fom_values = {}
-        context_metadata = {}
-        null_key = (_NULL_CONTEXT, _NULL_CONTEXT, frozenset())
-        context_metadata[null_key] = {
-            "name": _NULL_CONTEXT,
-            "def_name": _NULL_CONTEXT,
-            "vars": {},
-        }
-
-        # Iterate over files. We already know they exist
-        with lk.ReadTransaction(exp_lock):
-            for file, file_conf in files.items():
-
-                # Start with no active contexts in a file.
-                active_contexts = {}
-                logger.debug(f"Reading log file: {file}")
-
-                if not os.path.exists(file):
-                    logger.debug(
-                        f"Skipping analysis of non-existent file: {file}"
-                    )
-                    continue
-
-                per_file_crit_objs = [
-                    criteria_list.find_criteria(c)
-                    for c in file_conf["success_criteria"]
-                ]
-
-                with open(file, encoding="utf-8", errors="replace") as f:
-                    for line in f:
-                        new_per_file_crit_objs = []
-                        for crit_obj in per_file_crit_objs:
-                            if crit_obj.passed(line, self):
-                                crit_obj.mark_found()
-                            elif crit_obj.anti_matched(line):
-                                crit_obj.mark_anti_found()
-                            else:
-                                new_per_file_crit_objs.append(crit_obj)
-                        per_file_crit_objs = new_per_file_crit_objs
-
-                        # Iterate over contexts and add matched contexts to active_contexts
-                        for context, foms in file_conf["contexts"].items():
-                            if context != _NULL_CONTEXT:
-                                context_conf = f_defs[context]["definition"]
-                                if (
-                                    context_conf.get("pre_filter", "")
-                                    not in line
-                                ):
-                                    context_match = None
-                                else:
-                                    context_match = context_conf[
-                                        "regex"
-                                    ].match(line)
-
-                                if context_match:
-                                    context_name = format_context(
-                                        context_match,
-                                        context_conf["format"],
-                                    )
-                                    logger.debug(f"Line was: {line}")
-                                    logger.debug(
-                                        f" Context match {context} -- {context_name}"
-                                    )
-
-                                    context_vars = context_match.groupdict()
-                                    context_key = (
-                                        context_name,
-                                        context,
-                                        frozenset(context_vars.items()),
-                                    )
-
-                                    active_contexts[context] = context_key
-
-                                    if context_key not in fom_values:
-                                        fom_values[context_key] = {}
-                                        context_metadata[context_key] = {
-                                            "name": context_name,
-                                            "def_name": context,
-                                            "vars": context_vars,
-                                        }
-
-                            for fom in foms:
-                                fom_conf = f_defs[context]["foms"][fom]
-                                if fom_conf.get("pre_filter", "") not in line:
-                                    fom_match = None
-                                else:
-                                    fom_match = fom_conf["regex"].match(line)
-
-                                if fom_match:
-                                    fom_vars = fom_match.groupdict()
-                                    if (
-                                        fom_conf["fom_name_expanded"]
-                                        is not None
-                                    ):
-                                        fom_name = fom_conf[
-                                            "fom_name_expanded"
-                                        ]
-                                    else:
-                                        fom_name = self.expander.expand_var(
-                                            fom, extra_vars=fom_vars
-                                        )
-
-                                    if (
-                                        fom_conf["group"]
-                                        in fom_conf["regex"].groupindex
-                                    ):
-                                        logger.debug(
-                                            f" --- Matched fom {fom_name}"
-                                        )
-                                        fom_contexts = []
-                                        # if a FOM has contexts, check if each is active
-                                        if fom_conf["contexts"]:
-                                            for _ in fom_conf["contexts"]:
-                                                context_key = (
-                                                    active_contexts[context]
-                                                    if context
-                                                    in active_contexts
-                                                    else null_key
-                                                )
-                                                fom_contexts.append(
-                                                    context_key
-                                                )
-                                        else:
-                                            fom_contexts.append(null_key)
-
-                                        for fom_context in fom_contexts:
-                                            if fom_context not in fom_values:
-                                                fom_values[fom_context] = {}
-                                            fom_val = fom_match.group(
-                                                fom_conf["group"]
-                                            )
-                                            if fom_val is None:
-                                                continue
-                                            if (
-                                                fom_conf["units_expanded"]
-                                                is not None
-                                            ):
-                                                fom_unit = fom_conf["units"]
-                                            else:
-                                                fom_unit = (
-                                                    self.expander.expand_var(
-                                                        fom_conf["units"],
-                                                        extra_vars=fom_vars,
-                                                    )
-                                                )
-                                            fom_values[fom_context][
-                                                fom_name
-                                            ] = {
-                                                "value": fom_val,
-                                                "units": fom_unit,
-                                                "origin": fom_conf["origin"],
-                                                "origin_type": fom_conf[
-                                                    "origin_type"
-                                                ],
-                                                "fom_type": fom_conf[
-                                                    "fom_type"
-                                                ],
-                                            }
-        self.extract_inmem_foms(inmem_defs, fom_values, context_metadata)
-
-        # Test all non-file based success criteria
-        for criteria_obj, _ in criteria_list.all_criteria():
-            if criteria_obj.file is None:
-                if criteria_obj.passed(app_inst=self, fom_values=fom_values):
-                    criteria_obj.mark_found()
-
-        # If an app has no FOMs defined, don't fail it for that
-        success = (not f_defs and not inmem_defs) or False
-        for fom in fom_values.values():
-            for value in fom.values():
-                if (
-                    "origin_type" in value
-                    and value["origin_type"] == "application"
-                ):
-                    success = True
-        success = success and criteria_list.passed()
-
-        if success:
-            status = ExperimentStatus.SUCCESS
-        else:
-            preserved_terminal = {
-                ExperimentStatus.CANCELLED,
-                ExperimentStatus.TIMEOUT,
-                ExperimentStatus.FAILED,
-            }
-            current_status = self.get_status()
-            if current_status in preserved_terminal:
-                status = current_status
-            else:
-                status = ExperimentStatus.FAILED
-
-        # When workflow_manager is present, only use app_status when workflow is completed or
-        # unresolved.
-        if self.workflow_manager is not None:
-            wm_status = self.workflow_manager.get_status(workspace)
-            if not (
-                wm_status is None
-                or wm_status
-                in [ExperimentStatus.COMPLETE, ExperimentStatus.UNRESOLVED]
-            ):
-                status = wm_status
-
-        self.set_status(status)
-        self.result.finalize(workspace)
-
-        for criteria_obj, criteria_scope in criteria_list.all_criteria():
-            if criteria_obj.owner is not None:
-                criteria_name = (
-                    f"{criteria_obj.owner.scoped_name}::{criteria_obj.name}"
-                )
-            else:
-                criteria_name = (
-                    f"config::{criteria_scope}::{criteria_obj.name}"
-                )
-            if criteria_obj.ok():
-                self.result.success_criteria[criteria_name] = "PASSED"
-            else:
-                self.result.success_criteria[criteria_name] = "FAILED"
-
-        for context_key, fom_map in fom_values.items():
-            metadata = context_metadata[context_key]
-            context_map = {
-                "name": metadata["name"],
-                "foms": [],
-                "display_name": _get_context_display_name(metadata["name"]),
-                "context_def_name": metadata["def_name"],
-                "context_vars": metadata["vars"],
-            }
-
-            for fom_name, fom in fom_map.items():
-                fom_copy = fom.copy()
-                fom_copy["name"] = fom_name
-                context_map["foms"].append(fom_copy)
-
-            if metadata["name"] == _NULL_CONTEXT:
-                self.result.contexts.insert(0, context_map)
-            else:
-                self.result.contexts.append(context_map)
+        strategy_name = getattr(self, "analysis_strategy", "default")
+        strategy = ramble.analysis.get_strategy(strategy_name, self)
+        strategy(workspace)
 
     register_phase(
         "append_results_to_workspace",
